Re-organize unit tests
